refactor session card
